refactor: move benchmark harnesses into benchmarks package (by Lumen)
