+++ package.json
@@ -43,8 +43,8 @@
     "supabase:local:setup": "bash ./scripts/setup-local-supabase.sh",
     "test:integration:db:local": "bash ./scripts/test-integration-db-local.sh",
     "test:integration:runtime": "yarn workspace @inklabs/api test:integration:runtime",
-    "benchmark:memory-recall": "yarn workspace @inklabs/api benchmark:memory-recall",
-    "benchmark:bootstrap-relevance": "yarn workspace @inklabs/api benchmark:bootstrap-relevance",
+    "benchmark:memory-recall": "yarn workspace @inklabs/benchmarks benchmark:memory-recall",
+    "benchmark:bootstrap-relevance": "yarn workspace @inklabs/benchmarks benchmark:bootstrap-relevance",
     "backfill:memory-embeddings": "yarn workspace @inklabs/api backfill:memory-embeddings",
     "lint": "yarn workspaces foreach -A run lint",
     "type-check": "yarn workspaces foreach -A -t run type-check",
