+++ packages/api/package.json
@@ -25,8 +25,8 @@
     "test:coverage": "vitest run --coverage",
     "test:connection": "tsx src/test-connection.ts",
     "test:channels": "tsx src/test-channels.ts",
-    "benchmark:memory-recall": "tsx src/scripts/benchmark-memory-recall.ts",
-    "benchmark:bootstrap-relevance": "tsx src/scripts/benchmark-bootstrap-relevance.ts",
+    "benchmark:memory-recall": "yarn workspace @inklabs/benchmarks benchmark:memory-recall",
+    "benchmark:bootstrap-relevance": "yarn workspace @inklabs/benchmarks benchmark:bootstrap-relevance",
     "backfill:memory-embeddings": "tsx src/scripts/backfill-memory-embeddings.ts",
     "backfill:artifact-embeddings": "tsx src/scripts/backfill-artifact-embeddings.ts",
     "lint": "eslint src --ext .ts",
