+++ packages/benchmarks/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@inklabs/benchmarks",
+  "version": "0.5.0",
+  "private": true,
+  "license": "SEE LICENSE IN LICENSE",
+  "description": "Benchmark harnesses for Inkwell memory and runtime evaluation",
+  "scripts": {
+    "benchmark:memory-recall": "tsx src/benchmark-memory-recall.ts",
+    "benchmark:bootstrap-relevance": "tsx src/benchmark-bootstrap-relevance.ts",
+    "test": "vitest run --config vitest.config.ts"
+  },
+  "devDependencies": {
+    "tsx": "^4.20.6",
+    "vitest": "^4.0.18"
+  }
+}
