+++ packages/benchmarks/vitest.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    include: ['src/**/*.test.ts'],
+    exclude: ['node_modules', 'dist', '**/*.integration.test.ts'],
+  },
+});
